feat: implement robust error handling and refactor form services
- Created fetchWithRetry utility for reliable API calls with timeout and retry logic
- Refactored Salesforce form submission services for better error handling
- Improved notification services (Slack and OpenPhone) with more reliable delivery
- Enhanced form components with better validation and error feedback
- Standardized error handling patterns across all form submissions
- Added proper error logging for debugging and monitoring
