+++ utils/fetchWithRetry.ts
@@ -1,0 +1,31 @@
+export default async function fetchWithRetry(url: string, options: RequestInit, retries = 3, delay = 1000): Promise<Response> {
+    for (let i = 0; i < retries; i++) {
+        const controller = new AbortController();
+        const timeoutId = setTimeout(() => controller.abort(), 30000); // 30 seconds
+
+        try {
+            const response = await fetch(url, {
+                ...options,
+                signal: controller.signal,
+            });
+            clearTimeout(timeoutId);
+            if (!response.ok) {
+                throw new Error(`HTTP error: ${response.status}`);
+            }
+            return response;
+        } catch (error: unknown) {
+            clearTimeout(timeoutId);
+            if (error instanceof Error && error.name === 'AbortError') {
+                console.error(`Fetch aborted after 30s on attempt ${i + 1}`);
+            }
+            if (i < retries - 1) {
+                console.log(`Retrying fetch attempt ${i + 1} after ${delay}ms`);
+                await new Promise(resolve => setTimeout(resolve, delay));
+                delay *= 2; // Exponential backoff
+            } else {
+                throw error;
+            }
+        }
+    }
+    throw new Error('Max retries reached');
+}
